implemented datePicker

+++ src/screens/HomeScreen.js
@@ -4,27 +4,50 @@ import AsyncStorage from "@react-native-async-storage/async-storage";
 import moment from "moment";
 import { User, Entry } from "../classes";
 import constantStyles from "../constants/styles";
+import DateTimePicker from "@react-native-community/datetimepicker";
+import { Platform } from "react-native";
+
+const clearAsyncStorage = async () => {
+	const asyncStorageKeys = await AsyncStorage.getAllKeys();
+	if (asyncStorageKeys.length > 0) {
+		if (Platform.OS === "android") {
+			await AsyncStorage.clear();
+		}
+		if (Platform.OS === "ios") {
+			await AsyncStorage.multiRemove(asyncStorageKeys);
+		}
+	}
+};
 
 function HomeScreen() {
 	const [user, setUser] = useState();
 	const [weight, setWeight] = useState(0);
-	const [date, setDate] = useState(moment(new Date()).format("MM-DD-YYYY"));
+	const [date, setDate] = useState(new Date());
+	const [showDatePicker, setShowDatePicker] = useState(false);
 
 	const handleOnPress = async () => {
 		try {
-			let entry = new Entry(weight + 1, date);
+			let selectedWeight = weight + 1;
+			let selectedDate = moment(date).format("MM-DD-YYYY");
+			let entry = new Entry(selectedWeight, selectedDate);
 			user.addEntry(entry);
+			console.log(user);
 			setUser(user);
-			setWeight(weight + 1);
-			let data = JSON.stringify({ weight: weight + 1 });
+			setWeight(selectedWeight);
+			let data = JSON.stringify({ weight: selectedWeight });
 			await AsyncStorage.setItem("bodyScale_user", data);
 		} catch (err) {
 			console.log(err.message);
 		}
 	};
 
+	const handleDateChange = (e, selectedDate) => {
+		setDate(selectedDate);
+		setShowDatePicker(false);
+	};
+
 	const clearAsyncStorage = async () => {
-		await AsyncStorage.clear();
+		clearAsyncStorage();
 		setWeight(0);
 		let user = new User();
 		setUser(user);
@@ -39,7 +62,6 @@ function HomeScreen() {
 					setWeight(weight);
 				} else {
 					let user = new User();
-					console.log(user);
 					setUser = setUser(user);
 				}
 			} catch (err) {
@@ -52,16 +74,26 @@ function HomeScreen() {
 
 	return (
 		<View style={styles}>
-			<Text>HomeScreen</Text>
 			<TouchableOpacity onPress={handleOnPress}>
-				<Text>Click here to Log Count!!!!</Text>
+				<Text>{weight} lbs</Text>
+			</TouchableOpacity>
+			<TouchableOpacity onPress={() => setShowDatePicker(true)}>
+				<Text>{moment(date).format("MM-DD-YYYY")}</Text>
+				{showDatePicker && (
+					<DateTimePicker
+						style={{
+							marginBottom: "auto",
+						}}
+						value={date}
+						mode={"date"}
+						display={"spinner"}
+						onChange={handleDateChange}
+					/>
+				)}
 			</TouchableOpacity>
 			<TouchableOpacity onPress={clearAsyncStorage}>
 				<Text>Clear Async Storage</Text>
 			</TouchableOpacity>
-			<Text>{weight} lbs</Text>
-			{/* <Text>{user.entries}</Text> */}
-			<Text>Date: {date}</Text>
 		</View>
 	);
 }
@@ -70,6 +102,7 @@ const styles = StyleSheet.create({
 	...constantStyles,
 	alignItems: "center",
 	justifyContent: "center",
+	flexGrow: 1,
 });
 
 export default HomeScreen;
